fonts build freetype using emcc ports header files
